Feat: split voice functions into helper, added stop command for reading instructions
