+++ package.json
@@ -33,6 +33,7 @@
     "react-native-vector-icons": "^9.0.0",
     "react-native-web": "^0.17.5",
     "recoil": "^0.5.2",
+    "recoil-outside": "^0.1.1",
     "speech": "^0.2.1"
   },
   "devDependencies": {
